Split livery discovery and parsing.
