+++ dcs/liveries/liveryscanner.py
@@ -1,10 +1,7 @@
-import logging
 import os
 import re
-import zipfile
-from typing import Dict, Optional
+from typing import Dict
 
-import dcs.lua
 from dcs.installation import get_dcs_install_directory, get_dcs_saved_games_directory
 
 from .liberation import read_liberation_preferences
@@ -27,17 +24,6 @@ def safe_name(string: str) -> str:
     safe_name = re.sub(r"^(\d)", r"x_\1", safe_name)
     safe_name = re.sub(r"[\u0080-\uFFFF]", "_", safe_name)
     return safe_name
-
-
-def _attempt_read_from_filestream(filestream: bytes) -> Optional[str]:
-    encodes = ["utf-8", "ansi"]
-    for enc in encodes:
-        try:
-            code = filestream.decode(enc)
-        except UnicodeDecodeError:
-            continue
-        return code
-    return None
 
 
 class LiveryScanner:
@@ -66,104 +52,9 @@ class LiveryScanner:
         self.scan_custom_liveries(saved_games)
         return self.map
 
-    def scan_lua_code(self, luacode: str, path: str, unit: str) -> None:
-        """
-        Scans the given code (expecting contents from a description.lua file)
-        to extract the name of the livery together with the countries for which the livery is applicable.
-        If no name is found, it defaults to the folder-name like DCS would.
-        If no countries are found, it means the livery is valid for all countries.
-        Finally we also attempt to extract the order to sort liveries like DCS.
-        If no order is found, we use a default value of 0.
-
-        :param luacode: The contents of description.lua for the livery in question
-        :param path: The path of the livery in question
-        :param unit: The name of the unit as 'DCS type',
-            i.e. name of the unit inside the Liveries folder, e.g. 'A-10CII'
-        """
-        path_id = path.split("\\")[-1].replace(".zip", "")
-        if path_id == "placeholder":
-            return
-
-        # Some liveries files use variables in material names. Since the files read by
-        # the scanner can be arbitrarily changed by a game update (we don't want some
-        # liveries to be suddenly unparseable), and so far we aren't interested in the
-        # values of liveries that rely on undefined variables, just assume those are all
-        # the empty string and move on.
-        data = dcs.lua.loads(luacode, unknown_variable_lookup=lambda _: "")
-        livery_name = data.get("name", path_id)
-        countries_table = data.get("countries")
-        if countries_table is None:
-            countries = None
-        else:
-            countries = set(countries_table.values())
-        order = data.get("order", 0)
-
-        order = None if path_id == "default" else order
-        if order is not None and not isinstance(order, int):
-            try:
-                order = int(order)
-            except ValueError:
-                order = 0
-
-        livery = Livery(path_id, livery_name, order, countries)
+    def register_livery(self, unit: str, livery: Livery) -> None:
         self.map[unit].add(livery)
         setattr(self.map[unit], safe_name(livery.id), livery)
-
-    def scan_lua_description(self, path: str, unit: str) -> None:
-        """
-        Verifies a description file exists and reads its contents,
-        passing it to 'scan_lua_code'.
-
-        :param path: The path of the livery in question
-        :param unit: The name of the unit as 'DCS type',
-            i.e. name of the unit inside the Liveries folder, e.g. 'A-10CII'
-        """
-        description_path = os.path.join(path, "description.lua")
-        if os.path.exists(description_path):
-            # Known encodings used for description.lua files
-            with open(description_path, "rb") as file:
-                code = _attempt_read_from_filestream(file.read())
-                if code is None:
-                    logging.warning(f" Unknown encoding found in '{description_path}'")
-                    return
-
-                try:
-                    self.scan_lua_code(code, path, unit)
-                except (IndexError, SyntaxError):
-                    logging.getLogger("pydcs").exception(
-                        "Failed to parse Lua code in %s", description_path
-                    )
-                    print("Failed to parse Lua code in {}".format(description_path))
-                    raise
-
-    def scan_zip_file(self, path: str, unit: str) -> None:
-        """
-        Some liveries are zipped, this does the same as 'scan_lua_description',
-        except for the fact it needs to "open the zip" first.
-
-        :param path: The path of the livery in question
-        :param unit: The name of the unit as 'DCS type',
-            i.e. name of the unit inside the Liveries folder, e.g. 'A-10CII'
-        """
-        if os.path.exists(path):
-            with zipfile.ZipFile(path, "r") as zf:
-                if "description.lua" in zf.namelist():
-                    with zf.open("description.lua", "r") as file:
-                        code = _attempt_read_from_filestream(file.read())
-                        if code is None:
-                            logging.warning(
-                                f" Unknown encoding found in '{path}/description.lua'"
-                            )
-                            return
-
-                        try:
-                            self.scan_lua_code(code, path, unit)
-                        except (SyntaxError, IndexError):
-                            error_path = "!".join([path, "description.lua"])
-                            logging.getLogger("pydcs").exception(
-                                "Failed to parse Lua code in %s", error_path
-                            )
-                            raise
 
     def scan_liveries(self, path: str, campaign_path: bool = False) -> None:
         """
@@ -195,12 +86,10 @@ class LiveryScanner:
             if unit not in self.map:
                 self.map[unit] = LiverySet(unit)
                 setattr(self, safe_name(unit), self.map[unit])
-            for livery in os.listdir(liveries_path):
-                livery_path = os.path.join(liveries_path, livery)
-                if os.path.isdir(livery_path):
-                    self.scan_lua_description(livery_path, unit)
-                elif os.path.isfile(livery_path) and ".zip" in livery_path:
-                    self.scan_zip_file(livery_path, unit)
+            for livery_name in os.listdir(liveries_path):
+                livery = Livery.from_path(os.path.join(liveries_path, livery_name))
+                if livery is not None:
+                    self.register_livery(unit, livery)
 
     def scan_mods_path(self, path: str) -> None:
         """
